fix: Actions cell of admins table fixed.
All the roles are getted through the table index.tsx and passed to columns.tsx which is a function returning an array now. Thus in the columns the ActionsCell is getting all necessary props. The great thing about doing it like that is roles are fetched only one time when the table loads.